plot_2d_umap_anno

plot_2d_umap_anno

plot_2d_umap_anno(
  umap.fn,
  anno.df,
  dest.d = "./",
  umap.2d = NULL,
  meta.fields = NULL,
  show.label = FALSE,
  alpha = 0.65,
  cex = 0.25,
  save.format = c("png", "pdf", "both"),
  plot.height = 7,
  plot.width = 7,
  show.legend = FALSE
)

Arguments

umap.fn

path to umap coordinates. CSV file containing sample_id and umap x/y coordinates

anno.df

Sample annotations. The first column should be sample_id, and each annotation should have \_id, \_label, and \_color columns. Requires cluster_id which needs to be sequential in order of the dendrogram.

dest.d

path to save plots.

meta.fields

base name of variables to be represented as bargraphs below dendrogram. Annotation variables need to be represented as \_id, \_label, \_color in anno.

show.label

TRUE or FALSE. To show cluster label on top of plot.

alpha

level of transparency of dots. Value between 0 (fully transparent) and 1 (fully opaque)

cex

size of plotted points. Default = 0.25

save.format

figures can be saved as "png", "pdf" or "both"

plot.height

plot.height

plot.width

plot.width

show.legend

TRUE or FALSE. Whether to show legend with plot.

Examples