plot_3d_umap_anno

plot_3d_umap_anno

plot_3d_umap_anno(
  umap.fn,
  anno.df,
  cols = c("region", "cluster"),
  label_cols = list(NULL, "cluster_label"),
  win.dim = c(20, 40, 1500, 800),
  cex = 0.7,
  html.fn = NULL,
  dest.d = "./"
)

Arguments

umap.fn

path to umap coordinates. CSV file containing sample_id and umap x/y/z coordinates

anno.df

Sample annotations. The first column should be sample_id, and each annotation should have \_id, \_label, and \_color columns. Requires cluster_id which needs to be sequential in order of the dendrogram.

cols

base name of variables to be represented as bargraphs below dendrogram. Annotation variables need to be represented as \_id, \_label, \_color in anno.

label_cols

Which label on top of plot.

win.dim

dimensions of plot window

cex

size of plotted points. Default = 0.25

html.fn

name of html file to save figures can be saved as "png", "pdf" or "both"

dest.d

path to load data from and save plots to.

Examples

# First, load the example data
umap <- read.csv("data/rd_plot_example/3d.umap.sampled.csv")
#> Warning: cannot open file 'data/rd_plot_example/3d.umap.sampled.csv': No such file or directory
#> Error in file(file, "rt"): cannot open the connection
load("data/rd_plot_example/anno.df.3d.rda")
#> Warning: cannot open compressed file 'data/rd_plot_example/anno.df.3d.rda', probable reason 'No such file or directory'
#> Error in readChar(con, 5L, useBytes = TRUE): cannot open the connection

# Then, run the function with the loaded data
plots <- plot_3d_umap_anno(
  umap.fn = "3d.umap.sampled.csv",
  anno.df = anno.df,
  cols = c("platform", "cluster"),
  label_cols = list(NULL, NULL),
  html.fn = "3d.umap",
  dest.d = "data/rd_plot_example"
)
#> Error in fread(file.path(dest.d, umap.fn), header = TRUE): could not find function "fread"